feat(frontend): persist React Query cache offline; drop dead usersStore
Wrap the app in PersistQueryClientProvider backed by AsyncStorage so a
cold start can paint last-known catalog/library/profile data before the
network resolves. Only successful queries are dehydrated; a buster + 24h
maxAge discard stale or schema-changed snapshots. Guest/auth separation
is preserved by the existing query keys.

The Zustand usersStore was dead except for the UserEntity type consumed
by useProfileData (which already fetches via React Query). Relocate the
type into useProfileData and delete the store.
